Crate adbfbinlib

The adbfbinlib is a LIB and a facade to many modules that interact with the audiobookfinder program. In this file all crates are name before the modules which use them.

Modules

common
Common modules: common functionalities needed by the lib and also by the main program.

ctrl
The ctrl module should be the general controller of the program. Right now, most controlling is in the net module, and here is only a light facade to the tui messages.

data
The oldest module, the data module stores all the data needed to collect and also the search, dir algorithms

forwarder
The forwarder module shall act as wrapper for the ffi mangled lib.rs calls, which shall be very thin and spare there, but extensive here to prepare clean calls to adbfflutter.

net
The net module is resonsible for the network related parts, the mDNS registering, mDNS search, communication server and client. It also let's us startup and perform everything in yet one step.

shared
The shared module shall give binary and library high level functionality to be used by both the same way.

Structs

AdbflibError
A useless Error just for the Demo

Functions

file_count_good
the library interface for returning number of audio files found

get_own_peer
the library interface for returning own peer id

get_ui_messages
the library interface for returning ui messages