Module adbfbinlib::ctrl

The ctrl module should be the general controller of the program. Right now, most controlling is in the net module, and here is only a light facade to the tui messages.

Modules

tui
The TUI parts, greatly using Cursive showing table, the paths being searched, an alive for that, also the mDNS search performed and later status of the connection to the found clients.

webui
This is a webui about to replace the TUI, to be nice, better accessable, and new technology using websockets

Structs

Ctrl
The controller holds user interfaces as webui, tui. It currently creates and runs the user interfaces, distributes messages and sends out messages to be used somewhere else.

NetStatsMsg
NetStats message

UiClientPeer
Peer string identificators

Enums

CollectionPathAlive
alive Signal for path from collector or net search alive

Finisher

ForwardNetMsg
Forwarding net messages

InternalUiMsg
Internal messages inside UI

NetInfoMsg
Yet unimportant net messages todo: make it better!

Status
Turn on/off something

UiUpdateMsg
UI updating messages

Functions

try_open_browser