Module adbfbinlib::ctrl::webui::json

Structs

DirIn

DirOut

FinishPeer

PathData

PeerJson

StartData

Enums

AnimateData

NetData

RefreshData

WSJsonIn

WSJsonOut

Functions

REST_dirs
generate the REST dir paths to output json

convert_external_message

convert_internal_message
some internal shall fail because they are not supposed to be sent out, so it results in a correct MyJson or the attribute that shall not be send out but was tried to do ...

generate_init_data