Module adbfbinlib::data

The oldest module, the data module stores all the data needed to collect and also the search, dir algorithms

Modules

audio_info
Wraps bktree functionality, the container and defines audio info structs to be used

bktree

collection
The collection keeps and maintains all audio data.

ipc
IPC module will hold all massive (that is why IPC) internal messages which occur due to data collection, its start and its end.

tag_readers

Structs

IFInternalCollectionOutputData
Interface of what collection output data will return

Functions

publish_local_storage

search_in_single_path
The collection search! Searches on the file system a concrete, single path.