Module adbfbinlib::data::bktree

Structs

BKTree
BKTree structure that is used to store word like structures and perform "fuzzy" search on them to implement "do you mean" functionality on them. Can perform said search on any term that implements the distance trait. The default implementation is Osa distance.

BKTreeNode

Traits

Distance
This trait is used by the BKTree to determine the distance between 2 objects when fuzzy searching. An example of this for strings is the Levenshtein distance, Damerau-Levenshtein distance, Optimal string alignment distance, or a custom implementation.

Functions

osa_distance