Trait adbfbinlib::data::bktree::Distance

pub trait Distance {
    fn distance(&self, other: &Self) -> usize;
}

This trait is used by the BKTree to determine the distance between 2 objects when fuzzy searching. An example of this for strings is the Levenshtein distance, Damerau-Levenshtein distance, Optimal string alignment distance, or a custom implementation.

Required methods

fn distance(&self, other: &Self) -> usize

Used to determine the "distance" between two objects.

Loading content...