Function adbfbinlib::forwarder::create_net_runtime

fn create_net_runtime() -> (Receiver<UiUpdateMsg>, Sender<IPC>)

Opens a net thread and return ui message receiver and ipc message sender to be used as static instance for Dart-to-backend communication.