Module adbfbinlib::net

The net module is resonsible for the network related parts, the mDNS registering, mDNS search, communication server and client. It also let's us startup and perform everything in yet one step.

Modules

behavior
The net will be represented by a swarm as in libp2p https://docs.rs/libp2p/latest/libp2p/swarm/index.html.

sm
The StateChart for the server ... StateMachine type/struct is created inside here by macros I suppose, so using it needs to "reimport" this file/mod

sm_behaviour
Taken from dummy behaviour to have a layer of communication which reacts with the embedded state machine (and inner ui), also back to net services: currently kademlia, mdns https://docs.rs/libp2p/latest/libp2p/swarm/struct.DummyBehaviour.html

storage
wraps up "net storage" which is implementing the kademlia functionality.

subs
A folder/module for helper/sub modules

ui_data
a very small mod just for ui data send by net. It is important to not send all discovery blindly (e.g. duplicates)

Structs

Net
The Net component keeps control about everything from net.