Module adbfbinlib::net::sm

The StateChart for the server ... StateMachine type/struct is created inside here by macros I suppose, so using it needs to "reimport" this file/mod

Structs

AdbfStateChart

Go

NewPeerData

StartData

StateMachine
State machine structure definition.

UpdateData

Enums

Error
List of possible errors

Events
List of auto-generated events.

States
List of auto-generated states.

Traits

StateMachineContext
This trait outlines the guards and actions that need to be implemented for the state machine.