Module adbfbinlib::net::subs::peer_representation

Holds the peer representation, which is a hash of the PeerId which is also a hash or the public key inside. It's for convenience and smaller footprint.

Functions

peer_hash_to_string
Return a hex string hash reprentation of pure peer representation

peer_to_hash
Return PeerRepresentation directly from PeerId

peer_to_hash_string
Return a hex string hash representation of PeerId

Type Definitions

PeerRepresentation
Since peer lacks some functionality, PeerRepresentation is for convenience (shortening, Serialization)